Extract FixtureKit::RSpec::Declaration into its own file
