+++ lib/fixture_kit/rspec/declaration.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module FixtureKit
+  module RSpec
+    class Declaration
+      attr_reader :name
+
+      def initialize(name)
+        @name = name.to_s
+      end
+
+      def fixture_set
+        FixtureKit::FixtureRegistry.load_fixture(name)
+      end
+    end
+  end
+end
